zune_jpeg

Module components

This module exports a single struct to store information about JPEG image components

The data is extracted from a SOF header.

Structs

Components 🔒

Component Data from start of frame

Enums

ComponentID

Component ID’s

SampleRatios

Constants

MAX_SAMP_FACTOR 🔒

Type Aliases

UpSampler

Represents an up-sampler function, this function will be called to upsample a down-sampled image