zune_jpeg

Module misc

Miscellaneous stuff

Structs

Aligned16

Align data to a 16 byte boundary

Aligned32

Align data to a 32 byte boundary

Enums

SOFMarkers

Markers that identify different Start of Image markers They identify the type of encoding and whether the file use lossy(DCT) or lossless compression and whether we use Huffman or arithmetic coding schemes

Constants

START_OF_FRAME_BASE

Start of baseline DCT Huffman coding

START_OF_FRAME_EXT_AR

Start of extended sequential DCT arithmetic coding

START_OF_FRAME_EXT_SEQ

Start of another frame

START_OF_FRAME_LOS_SEQ

Start of Lossless sequential Huffman coding

START_OF_FRAME_LOS_SEQ_AR

Start of Lossless sequential Arithmetic coding

START_OF_FRAME_PROG_DCT

Start of progressive DCT encoding

START_OF_FRAME_PROG_DCT_AR

Start of Progressive DCT arithmetic coding

UN_ZIGZAG

Undo run length encoding of coefficients by placing them in natural order

Functions

calculate_padded_width

Calculate number of fill bytes added to the end of a JPEG image to fill the image

fill_default_mjpeg_tables

setup_component_params 🔒

Set up component parameters.